requirement.txt
matplotlib==3.3.4
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.18.5
pandas==1.2.1
scikit-learn==0.22.2.post1
sklearn==0.0
tensorflow==2.3.1
Pillow==2.2.2

